formatters: Refactor common code into Base class.
